feat: add unjustified absences with the departure date of the intern.
In this code I have added the functionality of being able to add the unjustified absences of the intern with the departure date. On the administrator page I added it in summary, interns and intern-profile. I added two files, one to be able to count the absences in a function and another to update the departure date.